pyproject.toml
[tool.poetry]
name = "obsidian-plugins-from-github"
version = "0.1.0"
description = "Gets a list of obsidian plugins from GitHub."
authors = ["João Luís <jolitp@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
markdown = "^3.7"
beautifulsoup4 = "^4.12.3"
rich = "^13.8.0"
urllib3 = "^2.2.2"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
black = "^24.8.0"
isort = "^5.13.2"
taskipy = "^1.13.0"


[tool.poetry.group.doc.dependencies]
mkdocs-material = "^9.5.34"
mkdocstrings = "^0.26.1"
mkdocstrings-python = "^1.11.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = "."
addopts = "--doctest-modules"

[tool.isort]
profile = "black"

[tool.taskipy.tasks]
lint = "black --check --diff . && isort --check --diff ."
docs = "mkdocs serve"
pre_test = "task lint"
test = "pytest -s -x --cov=obsidian_plugins_from_github -vv"
post_test = "coverage html"
run = "poetry run python main.py"